// apps/web/lib/garden-planner/enums/cropCode.ts
enum CropCode {
    None = 'N',
    Tomato = 'T',
    Potato = 'P',
    Rice = 'R',
    Wheat = 'W',
    Carrot = 'C',
    Onion = 'O',
    Cotton = 'Co',
    Blueberry = 'B',
    Apple = 'A',
    Corn = 'Cr',
    SpicyPepper = 'S',
    NapaCabbage = 'Cb',
    BokChoy = 'Bk',
    RockhopperPumpkin = 'Pm',
    BatterflyBean = 'Bt'
}

export default CropCode 